// apps/web/src/api/generated/model/unavailableConnectionDtoReason.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */

export type UnavailableConnectionDtoReason = typeof UnavailableConnectionDtoReason[keyof typeof UnavailableConnectionDtoReason];


export const UnavailableConnectionDtoReason = {
  timeout: 'timeout',
  unreachable: 'unreachable',
  unauthorized: 'unauthorized',
  bad_response: 'bad_response',
  blocked_host: 'blocked_host',
} as const;